repair: damage from int to double

+++ src/main/java/at/fhtw/swen/mctg/core/cards/Spell.java
@@ -4,7 +4,7 @@ import at.fhtw.swen.mctg.model.Card;
 
 public class Spell extends Card {
 
-    public Spell(int damage, Element type) {
+    public Spell(double damage, Element type) {
         super(damage, type);
     }
     @Override
