+++ src/main/java/at/fhtw/swen/mctg/core/cards/monsters/Dragon.java
@@ -3,7 +3,7 @@ package at.fhtw.swen.mctg.core.cards.monsters;
 import at.fhtw.swen.mctg.core.cards.Monster;
 
 public class Dragon extends Monster {
-    public Dragon(int damage, Element type) {
+    public Dragon(double damage, Element type) {
         super(damage, type);
     }
